Fix compiler warning by casting time_t to long.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@18774 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ socket/metaserver.c
@@ -440,7 +440,7 @@ static void metaserver2_updates(void) {
                  CURLFORM_COPYCONTENTS, buf,
                  CURLFORM_END);
 
-    snprintf(buf, sizeof(buf), "%ld", metaserver2_updateinfo.uptime);
+    snprintf(buf, sizeof(buf), "%ld", (long)metaserver2_updateinfo.uptime);
     curl_formadd(&formpost, &lastptr,
                  CURLFORM_COPYNAME, "uptime",
                  CURLFORM_COPYCONTENTS, buf,
